move get tenancy into sql tenancy usecase
