+++ config/routes.rb
@@ -7,6 +7,7 @@ Rails.application.routes.draw do
     post '/users/find-or-create', to: 'users#create'
     patch '/tenancies/:tenancy_ref', to: 'tenancies#update'
     get '/tenancies/:tenancy_ref/pause', to: 'tenancies#pause'
+    get '/tenancies/:tenancy_ref', to: 'tenancies#show'
     post '/tenancies/:tenancy_ref/action_diary', to: 'action_diary#create'
     post '/messages/send_sms', to: 'messages#send_sms'
     post '/messages/send_email', to: 'messages#send_email'
