+++ lib/hackney/income/sql_tenancy_case_gateway.rb
@@ -30,6 +30,17 @@ module Hackney
         tenancy.assigned_user_id
       end
 
+      def get_tenancy(tenancy_ref:)
+        tenancy = GatewayModel.find_by(tenancy_ref: tenancy_ref)
+        if tenancy.nil?
+          Rails.logger.error("Failed to retrieve tenancy with tenancy_ref: '#{tenancy_ref}' ")
+          raise TenancyNotFoundError, "Unable to get tenancy: #{tenancy_ref} - tenancy not found."
+        end
+        tenancy
+      end
+
+      class TenancyNotFoundError < StandardError; end
+
       private
 
       def next_available_user_for(band:)
